Support in OrderContents#update_cart for Multiple Line-Items (#3)
* Fixes Context Nesting in OrderContents Spec

* Pin Capybara Version in Gemspec

The test suite breaks if Capybara 3 is used. Since support for Capybara 3 is
out-of-scope for this effort, I've pinned the Capybara version to ~> 2 instead.

* Improves OrderContents#update_cart

Because Spree::Order accepts nested attributes for line-items, it's possible to
add _multiple_ line-items to an order simultaneously via
OrderContents#update_cart. However, the virtual gift-cards decoration of this
method doesn't support this usage. Now it does!

* Pins 'mysql' and 'pg' to Known-Working Versions